requirements.txt
# Libraries Essentielles pour commencer
pandas==2.1.0
numpy==1.25.0
scikit-learn==1.3.0
plotly==5.15.0
matplotlib==3.7.0
seaborn==0.12.0

# Base de données
sqlalchemy==2.0.20
psycopg2-binary==2.9.7

# API et Web
fastapi==0.103.0
uvicorn==0.23.0
requests==2.31.0

# Utilitaires
python-dotenv==1.0.0
loguru==0.7.0